remove CBC w/HMAC ciphers
this is a performance improvement not a security oversight. these ciphers were indeed implemented securely as deployed in nuOS, soon after the Lucky13 attack was discovered and published. the secure implementation of CBC w/HMAC in TLSv1.2 made a performance tradeoff which is a suboptimal workaround and i think sufficient time has passed to disable them altogether