set volmode on parent dataset if it must be created. TODO: parameterize the dataset and volmode; warn if leaf volmode != dataset volmode
i added this because we all sometimes want to bypass geom and cam by with a dd straight into or out of a vol in img. for now im continuing to explictly set the leaf volmode although it seems redundant. TODO: the gui should present a different icon or color or border or whathaveyou according to volmode and somehow make any vol.volmode != ds.volmode stick out like a sore thumb