support decompression (-Z); first hack at outputting encrypted streams from encrypted datasets (the default; use -E for previous behavior)
the current problem is that the inheritance or reference of a non-identical encryptionroot property becomes broken in the restoration, leading to a logistical nightmare visa vis `zfs key-load` and `zfs change-key`. however, this presently works in a pinch (so long as you are well aware of the explosion into an encryptionroot per each encrypted dataset--i.e. dont forget youre left with stale keys after modifying encryption parameters on only the dataset you expected was the encryptionroot). so, i'm committing this work in progress because i have to shift gears onto other work for a few days